@@ -9,10 +9,11 @@
     "API"
   ],
   "require": {
-    "php": ">=5.6",
+    "php": "^7.1",
     "ext-mbstring": "*"
   },
   "require-dev": {
+    "doctrine/coding-standard": "^4.0",
     "phpunit/phpunit": "^4.8",
     "psr/http-message": "^1.0"
   },
@@ -35,5 +36,8 @@
   "suggest": {
     "react/promise": "To leverage async resolving on React PHP platform",
     "psr/http-message": "To use standard GraphQL server"
+  },
+  "scripts": {
+    "lint" : "vendor/bin/phpcs"
   }
 }
