@@ -1,5 +1,11 @@
 <?php
+
+declare(strict_types=1);
+
 namespace GraphQL\Error;
+
+use const E_USER_WARNING;
+use function trigger_error;
 
 /**
  * Encapsulates warnings produced by the library.
@@ -9,27 +15,29 @@ namespace GraphQL\Error;
  */
 final class Warning
 {
-    const WARNING_ASSIGN = 2;
-    const WARNING_CONFIG = 4;
-    const WARNING_FULL_SCHEMA_SCAN = 8;
+    const WARNING_ASSIGN             = 2;
+    const WARNING_CONFIG             = 4;
+    const WARNING_FULL_SCHEMA_SCAN   = 8;
     const WARNING_CONFIG_DEPRECATION = 16;
-    const WARNING_NOT_A_TYPE = 32;
-    const ALL = 63;
+    const WARNING_NOT_A_TYPE         = 32;
+    const ALL                        = 63;
 
-    static $enableWarnings = self::ALL;
+    /** @var int */
+    private static $enableWarnings = self::ALL;
 
-    static $warned = [];
+    /** @var mixed[] */
+    private static $warned = [];
 
-    static private $warningHandler;
+    /** @var callable|null */
+    private static $warningHandler;
 
     /**
      * Sets warning handler which can intercept all system warnings.
      * When not set, trigger_error() is used to notify about warnings.
      *
      * @api
-     * @param callable|null $warningHandler
      */
-    public static function setWarningHandler(callable $warningHandler = null)
+    public static function setWarningHandler(?callable $warningHandler = null)
     {
         self::$warningHandler = $warningHandler;
     }
@@ -45,14 +53,15 @@ final class Warning
      * @api
      * @param bool|int $suppress
      */
-    static function suppress($suppress = true)
+    public static function suppress($suppress = true)
     {
-        if (true === $suppress) {
+        if ($suppress === true) {
             self::$enableWarnings = 0;
-        } else if (false === $suppress) {
+        } elseif ($suppress === false) {
             self::$enableWarnings = self::ALL;
         } else {
             $suppress = (int) $suppress;
+
             self::$enableWarnings &= ~$suppress;
         }
     }
@@ -70,33 +79,34 @@ final class Warning
      */
     public static function enable($enable = true)
     {
-        if (true === $enable) {
+        if ($enable === true) {
             self::$enableWarnings = self::ALL;
-        } else if (false === $enable) {
+        } elseif ($enable === false) {
             self::$enableWarnings = 0;
         } else {
             $enable = (int) $enable;
+
             self::$enableWarnings |= $enable;
         }
     }
 
-    static function warnOnce($errorMessage, $warningId, $messageLevel = null)
+    public static function warnOnce($errorMessage, $warningId, $messageLevel = null)
     {
         if (self::$warningHandler) {
             $fn = self::$warningHandler;
             $fn($errorMessage, $warningId);
-        } else if ((self::$enableWarnings & $warningId) > 0 && !isset(self::$warned[$warningId])) {
+        } elseif ((self::$enableWarnings & $warningId) > 0 && ! isset(self::$warned[$warningId])) {
             self::$warned[$warningId] = true;
             trigger_error($errorMessage, $messageLevel ?: E_USER_WARNING);
         }
     }
 
-    static function warn($errorMessage, $warningId, $messageLevel = null)
+    public static function warn($errorMessage, $warningId, $messageLevel = null)
     {
         if (self::$warningHandler) {
             $fn = self::$warningHandler;
             $fn($errorMessage, $warningId);
-        } else if ((self::$enableWarnings & $warningId) > 0) {
+        } elseif ((self::$enableWarnings & $warningId) > 0) {
             trigger_error($errorMessage, $messageLevel ?: E_USER_WARNING);
         }
     }
