@@ -1,10 +1,18 @@
 <?php
+
+declare(strict_types=1);
+
 namespace GraphQL\Error;
 
 use GraphQL\Language\AST\Node;
 use GraphQL\Language\Source;
 use GraphQL\Language\SourceLocation;
 use GraphQL\Utils\Utils;
+use function array_filter;
+use function array_map;
+use function array_merge;
+use function is_array;
+use function iterator_to_array;
 
 /**
  * Describes an Error found during the parse, validate, or
@@ -22,7 +30,7 @@ use GraphQL\Utils\Utils;
  */
 class Error extends \Exception implements \JsonSerializable, ClientAware
 {
-    const CATEGORY_GRAPHQL = 'graphql';
+    const CATEGORY_GRAPHQL  = 'graphql';
     const CATEGORY_INTERNAL = 'internal';
 
     /**
@@ -32,23 +40,21 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      */
     public $message;
 
-    /**
-     * @var SourceLocation[]
-     */
+    /** @var SourceLocation[] */
     private $locations;
 
     /**
      * An array describing the JSON-path into the execution response which
      * corresponds to this error. Only included for errors during execution.
      *
-     * @var array
+     * @var mixed[]|null
      */
     public $path;
 
     /**
      * An array of GraphQL AST Nodes corresponding to this error.
      *
-     * @var array
+     * @var Node[]|null
      */
     public $nodes;
 
@@ -62,34 +68,74 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      */
     private $source;
 
-    /**
-     * @var array
-     */
+    /** @var int[]|null */
     private $positions;
 
-    /**
-     * @var bool
-     */
+    /** @var bool */
     private $isClientSafe;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $category;
 
-    /**
-     * @var array
-     */
+    /** @var mixed[]|null */
     protected $extensions;
+
+    /**
+     * @param string       $message
+     * @param Node[]|null  $nodes
+     * @param mixed[]|null $positions
+     * @param mixed[]|null $path
+     * @param \Throwable   $previous
+     * @param mixed[]      $extensions
+     */
+    public function __construct(
+        $message,
+        $nodes = null,
+        ?Source $source = null,
+        $positions = null,
+        $path = null,
+        $previous = null,
+        array $extensions = []
+    ) {
+        parent::__construct($message, 0, $previous);
+
+        // Compute list of blame nodes.
+        if ($nodes instanceof \Traversable) {
+            $nodes = iterator_to_array($nodes);
+        } elseif ($nodes && ! is_array($nodes)) {
+            $nodes = [$nodes];
+        }
+
+        $this->nodes      = $nodes;
+        $this->source     = $source;
+        $this->positions  = $positions;
+        $this->path       = $path;
+        $this->extensions = $extensions ?: (
+        $previous && $previous instanceof self
+            ? $previous->extensions
+            : []
+        );
+
+        if ($previous instanceof ClientAware) {
+            $this->isClientSafe = $previous->isClientSafe();
+            $this->category     = $previous->getCategory() ?: static::CATEGORY_INTERNAL;
+        } elseif ($previous) {
+            $this->isClientSafe = false;
+            $this->category     = static::CATEGORY_INTERNAL;
+        } else {
+            $this->isClientSafe = true;
+            $this->category     = static::CATEGORY_GRAPHQL;
+        }
+    }
 
     /**
      * Given an arbitrary Error, presumably thrown while attempting to execute a
      * GraphQL operation, produce a new GraphQLError aware of the location in the
      * document responsible for the original Error.
      *
-     * @param $error
-     * @param array|null $nodes
-     * @param array|null $path
+     * @param mixed        $error
+     * @param Node[]|null  $nodes
+     * @param mixed[]|null $path
      * @return Error
      */
     public static function createLocatedError($error, $nodes = null, $path = null)
@@ -99,22 +145,22 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
                 return $error;
             } else {
                 $nodes = $nodes ?: $error->nodes;
-                $path = $path ?: $error->path;
+                $path  = $path ?: $error->path;
             }
         }
 
-        $source = $positions = $originalError = null;
+        $source     = $positions = $originalError = null;
         $extensions = [];
 
         if ($error instanceof self) {
-            $message = $error->getMessage();
+            $message       = $error->getMessage();
             $originalError = $error;
-            $nodes = $error->nodes ?: $nodes;
-            $source = $error->source;
-            $positions = $error->positions;
-            $extensions = $error->extensions;
-        } else if ($error instanceof \Exception || $error instanceof \Throwable) {
-            $message = $error->getMessage();
+            $nodes         = $error->nodes ?: $nodes;
+            $source        = $error->source;
+            $positions     = $error->positions;
+            $extensions    = $error->extensions;
+        } elseif ($error instanceof \Exception || $error instanceof \Throwable) {
+            $message       = $error->getMessage();
             $originalError = $error;
         } else {
             $message = (string) $error;
@@ -131,64 +177,12 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
         );
     }
 
-
     /**
-     * @param Error $error
-     * @return array
+     * @return mixed[]
      */
     public static function formatError(Error $error)
     {
         return $error->toSerializableArray();
-    }
-
-    /**
-     * @param string $message
-     * @param array|Node|null $nodes
-     * @param Source $source
-     * @param array|null $positions
-     * @param array|null $path
-     * @param \Throwable $previous
-     * @param array $extensions
-     */
-    public function __construct(
-        $message,
-        $nodes = null,
-        Source $source = null,
-        $positions = null,
-        $path = null,
-        $previous = null,
-        array $extensions = []
-    )
-    {
-        parent::__construct($message, 0, $previous);
-
-        // Compute list of blame nodes.
-        if ($nodes instanceof \Traversable) {
-            $nodes = iterator_to_array($nodes);
-        } else if ($nodes && !is_array($nodes)) {
-            $nodes = [$nodes];
-        }
-
-        $this->nodes = $nodes;
-        $this->source = $source;
-        $this->positions = $positions;
-        $this->path = $path;
-        $this->extensions = $extensions ?: (
-            $previous && $previous instanceof self
-                ? $previous->extensions
-                : []
-        );
-
-        if ($previous instanceof ClientAware) {
-            $this->isClientSafe = $previous->isClientSafe();
-            $this->category = $previous->getCategory() ?: static::CATEGORY_INTERNAL;
-        } else if ($previous) {
-            $this->isClientSafe = false;
-            $this->category = static::CATEGORY_INTERNAL;
-        } else {
-            $this->isClientSafe = true;
-            $this->category = static::CATEGORY_GRAPHQL;
-        }
     }
 
     /**
@@ -212,29 +206,36 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      */
     public function getSource()
     {
-        if (null === $this->source) {
-            if (!empty($this->nodes[0]) && !empty($this->nodes[0]->loc)) {
+        if ($this->source === null) {
+            if (! empty($this->nodes[0]) && ! empty($this->nodes[0]->loc)) {
                 $this->source = $this->nodes[0]->loc->source;
             }
         }
+
         return $this->source;
     }
 
     /**
-     * @return array
+     * @return int[]
      */
     public function getPositions()
     {
-        if (null === $this->positions) {
-            if (!empty($this->nodes)) {
-                $positions = array_map(function($node) {
+        if ($this->positions === null && ! empty($this->nodes)) {
+            $positions = array_map(
+                function ($node) {
                     return isset($node->loc) ? $node->loc->start : null;
-                }, $this->nodes);
-                $this->positions = array_filter($positions, function($p) {
+                },
+                $this->nodes
+            );
+
+            $this->positions = array_filter(
+                $positions,
+                function ($p) {
                     return $p !== null;
-                });
-            }
+                }
+            );
         }
+
         return $this->positions;
     }
 
@@ -254,21 +255,29 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      */
     public function getLocations()
     {
-        if (null === $this->locations) {
+        if ($this->locations === null) {
             $positions = $this->getPositions();
-            $source = $this->getSource();
-            $nodes = $this->nodes;
+            $source    = $this->getSource();
+            $nodes     = $this->nodes;
 
             if ($positions && $source) {
-                $this->locations = array_map(function ($pos) use ($source) {
-                    return $source->getLocation($pos);
-                }, $positions);
-            } else if ($nodes) {
-                $this->locations = array_filter(array_map(function ($node) {
-                    if ($node->loc) {
-                        return $node->loc->source->getLocation($node->loc->start);
-                    }
-                }, $nodes));
+                $this->locations = array_map(
+                    function ($pos) use ($source) {
+                        return $source->getLocation($pos);
+                    },
+                    $positions
+                );
+            } elseif ($nodes) {
+                $this->locations = array_filter(
+                    array_map(
+                        function ($node) {
+                            if ($node->loc) {
+                                return $node->loc->source->getLocation($node->loc->start);
+                            }
+                        },
+                        $nodes
+                    )
+                );
             } else {
                 $this->locations = [];
             }
@@ -278,7 +287,7 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
     }
 
     /**
-     * @return array|Node[]|null
+     * @return Node[]|null
      */
     public function getNodes()
     {
@@ -290,7 +299,7 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      * Only included for execution errors.
      *
      * @api
-     * @return array|null
+     * @return mixed[]|null
      */
     public function getPath()
     {
@@ -298,7 +307,7 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
     }
 
     /**
-     * @return array
+     * @return mixed[]
      */
     public function getExtensions()
     {
@@ -309,26 +318,29 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      * Returns array representation of error suitable for serialization
      *
      * @deprecated Use FormattedError::createFromException() instead
-     * @return array
+     * @return mixed[]
      */
     public function toSerializableArray()
     {
         $arr = [
-            'message' => $this->getMessage()
+            'message' => $this->getMessage(),
         ];
 
         if ($this->getExtensions()) {
             $arr = array_merge($this->getExtensions(), $arr);
         }
 
-        $locations = Utils::map($this->getLocations(), function(SourceLocation $loc) {
-            return $loc->toSerializableArray();
-        });
+        $locations = Utils::map(
+            $this->getLocations(),
+            function (SourceLocation $loc) {
+                return $loc->toSerializableArray();
+            }
+        );
 
-        if (!empty($locations)) {
+        if (! empty($locations)) {
             $arr['locations'] = $locations;
         }
-        if (!empty($this->path)) {
+        if (! empty($this->path)) {
             $arr['path'] = $this->path;
         }
 
@@ -340,9 +352,8 @@ class Error extends \Exception implements \JsonSerializable, ClientAware
      * @link http://php.net/manual/en/jsonserializable.jsonserialize.php
      * @return mixed data which can be serialized by <b>json_encode</b>,
      * which is a value of any type other than a resource.
-     * @since 5.4.0
      */
-    function jsonSerialize()
+    public function jsonSerialize()
     {
         return $this->toSerializableArray();
     }
