@@ -1,12 +1,13 @@
 <?php
+
+declare(strict_types=1);
+
 namespace GraphQL\Error;
 
 /**
  * Class UserError
  *
  * Error caused by actions of GraphQL clients. Can be safely displayed to a client...
- *
- * @package GraphQL\Error
  */
 class UserError extends \RuntimeException implements ClientAware
 {
